fastapi/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart
sqlalchemy
python-dotenv
pydantic
email-validator
PyMySQL
brevo_python
httpx>=0.27.0
stripe>=12.0.0
websockets
fastapi_mail
pillow